[testloop] Convert TestLoopAsyncComputationSpawner to use CallbackEvent
