+++ core/async/src/test_loop/futures.rs
@@ -1,5 +1,5 @@
 use super::{delay_sender::DelaySender, event_handler::LoopEventHandler, TestLoop};
-use crate::futures::{AsyncComputationSpawner, DelayedActionRunner};
+use crate::futures::DelayedActionRunner;
 use crate::test_loop::event_handler::TryIntoOrSelf;
 use crate::time::Duration;
 use crate::{futures::FutureSpawner, messaging::CanSend};
@@ -187,37 +187,4 @@ impl<Data: 'static, Event: Debug + Send + 'static> TestLoop<Vec<Data>, (usize, E
             .for_index(idx),
         );
     }
-}
-
-/// An event that represents async computation. See async_computation_spawner() in DelaySender.
-pub struct TestLoopAsyncComputationEvent {
-    name: String,
-    f: Box<dyn FnOnce() + Send>,
-}
-
-impl Debug for TestLoopAsyncComputationEvent {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_tuple("AsyncComputation").field(&self.name).finish()
-    }
-}
-
-/// AsyncComputationSpawner that spawns the computation in the TestLoop.
-pub struct TestLoopAsyncComputationSpawner {
-    pub(crate) sender: DelaySender<TestLoopAsyncComputationEvent>,
-    pub(crate) artificial_delay: Box<dyn Fn(&str) -> Duration + Send + Sync>,
-}
-
-impl AsyncComputationSpawner for TestLoopAsyncComputationSpawner {
-    fn spawn_boxed(&self, name: &str, f: Box<dyn FnOnce() + Send>) {
-        self.sender.send_with_delay(
-            TestLoopAsyncComputationEvent { name: name.to_string(), f },
-            (self.artificial_delay)(name),
-        );
-    }
-}
-
-pub fn drive_async_computations() -> LoopEventHandler<(), TestLoopAsyncComputationEvent> {
-    LoopEventHandler::new_simple(|event: TestLoopAsyncComputationEvent, _data: &mut ()| {
-        (event.f)();
-    })
 }
